Renamed PDFExporter -> downloadPageAsPDF
