+++ packages/tsutils/src/index.ts
@@ -1,3 +1,3 @@
 export * from './validation';
 export * from './types';
-export * from './PDFExporter';
+export * from './downloadPageAsPDF';
